Add files via upload
The BOT.Pic contains a method that gets all the tags of the pictures in a website. the Download Pic.py uses beautiful soup and request to web scrape and download all the pictures to a selected folder. The selenium part then post all those pictures to instagram and this happens every 24 hours.